Build OpenSSL 3.0.7
I added the source code for OpenSSL to the repository as a Git
submodule. I checked out the source code for the 3.0.7 release of
OpenSSL.

I created the build.sh program and ios-and-catalyst.conf configuration
file. build.sh will build the OpenSSL libraries and will create the
XCFrameworks for libcrypto and libssl from the libraries.
ios-and-catalyst.conf contains build configuration settings for the
OpenSSL build system that are used to produce the iOS and macOS Catalyst
versions of the OpenSSL libraries.