OpenSSL 3.0.7 support (#1)
I added the source code for OpenSSL to the repository as a Git
submodule. I checked out the source code for the 3.0.7 release of
OpenSSL.

I created the build.sh program and ios-and-catalyst.conf configuration
file. build.sh will build the OpenSSL libraries and will create the
XCFrameworks for libcrypto and libssl from the libraries.
ios-and-catalyst.conf contains build configuration settings for the
OpenSSL build system that are used to produce the iOS and macOS Catalyst
versions of the OpenSSL libraries.

I created the Swift package to wrap libcrypto and libssl. I created the
COpenSSL module to export the OpenSSL code to Swift. I created the
OpenSSL library to expose a Swift-compatible OpenSSL API to Swift
applications or other frameworks.

I created .github/release.yml and configured the rules that GitHub uses
to automatically generate the release notes for a new product release.